black_region_detector/requirements.txt
opencv-python>=4.8.0
numpy>=1.24.0
pyserial>=3.5
Pillow>=10.0.0
PyYAML>=6.0

# ROS2 support (optional, if you need ROS)
# rclpy>=3.3.0
# geometry-msgs>=4.2.0

